refactor: publish script and ci workflow
