+++ .github/workflows/publish.yml
@@ -1,40 +1,46 @@
-name: Publish CI
+name: Publish packages
 
 on:
   push:
-    tags: [v*]
-    branches: [main]
-  pull_request:
-    branches: [main]
+    tags:
+      - v*
 jobs:
-  publish:
+  publish-NPM:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
 
-      - name: Use Node.js 16.x
-        uses: actions/setup-node@v1
+      - name: Setup Node.js environment
+        uses: actions/setup-node@v3.2.0
         with:
-          node-version: '16.x'
+          node-version: 16.x
 
-#       - name: Get yarn cache directory path
-#         id: yarn-cache-dir-path
-#         run: echo "::set-output name=dir::$(yarn cache dir)"
+      - name: Setup PNPM
+        uses: pnpm/action-setup@v2.2.1
+        with:
+          run_install: false
 
-#       - uses: actions/cache@v2
-#         id: yarn-cache
-#         with:
-#           path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
-#           key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+      - name: Get pnpm store directory
+        id: pnpm-cache
+        run: |
+          echo "::set-output name=pnpm_cache_dir::$(pnpm store path)"
 
-      - name: Install
+      - uses: actions/cache@v3
+        name: Setup pnpm cache
+        with:
+          path: ${{ steps.pnpm-cache.outputs.pnpm_cache_dir }}
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            ${{ runner.os }}-pnpm-store-
+
+      - name: Install dependencies
         run: pnpm install
 
       - name: Build
         run: pnpm build
 
       - name: Publish packages to NPM
-        uses: JS-DevTools/npm-publish@v1
-        with:
-          package: './dist/package.json'
-          token: ${{ secrets.NPM_TOKEN }}
+        run: pnpm publish:ci
+        env:
+          NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
+          NODE_OPTIONS: --max-old-space-size=6144
