+++ package.json
@@ -12,7 +12,8 @@
     "build:types": "tsc",
     "format": "prettier --write \"src/**/*.{ts,vue,tsx}\" \"test/**/*.{ts,vue,tsx}\" --vue-indent-script-and-styl",
     "format:dist": "prettier --write \"dist/**/*.{js,d.ts}\"",
-    "publish": "esno ./scripts/publish.ts",
+    "push-tag": "test && esno ./scripts/push-tag.ts",
+    "publish:ci": "esno ./scripts/publish.ts",
     "test": "vitest -w",
     "test:ui": "vitest -w --ui",
     "coverage": "vitest run --coverage",
@@ -47,7 +48,6 @@
     "semver": "^7.3.7",
     "typescript": "^4.5.4",
     "vite": "^2.9.7",
-    "vite-plugin-dts": "^1.1.1",
     "vitepress": "^0.22.4",
     "vitest": "^0.12.0",
     "vue": "^2.6.14 || >=3.0.0",
