+++ packages/hooks/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@vueposu/hooks",
+  "name": "vueposu",
   "version": "0.4.0",
   "description": "A hooks library based on Vue Composition-API",
   "repository": {
